Fix how the timestamps of add-source packages are updated.
This is to address #1375.

Previously they were updated in the post install phase, which
would mean that the sources did not have their timestamps
updated unless the package being installed completed successfully.

This change moves the updating of the timestamps to the end of
the install phase, so that the add-source packages that are
successfully installed get their timestamps updated regardless of
the overall outcome of the installation.