Merge dependencies more effectively in D.Solver.Modular.IndexConversion.
This commit is a more thorough fix for the problem that
24856040aea4e40a63b040fb61ef887070f06974 attempted to fix.  24856040ae combined
all dependencies on the same package within the same build-depends or
build-tool-depends field, in order to avoid performance problems in the solver
QuickCheck tests.  That wasn't enough to prevent the solver from creating many
duplicate dependencies when converting the PackageDescription to the
solver-specific format, though.  When a package contained many conditionals that
each contained the same dependency in both branches, the solver lifted those
dependencies out of the conditionals without deduplicating them.

This commit moves the step that merges dependencies after the step that lifts
common dependencies out of conditionals.