Don't ignore goal order parameter when generating summarized solver log.
After #5012, the solver generated the summarized log by using a goal order that
preferred goals from the final conflict set.  It completely ignored the original
goal order parameter.  This caused a failure in one of the unit tests that set a
goal order and then checked the summarized log, when the solver's goal order
changed for other reasons (See #5012).

This commit combines the two goal orders, with the goal order that prefers goals
from the conflict set taking precedence over the original goal order.