Enforce the backjump limit before creating the solver log.
Previously, the solver enforced the backjump limit by counting 'Backjump'
messages during log processing.  This commit instead enforces the backjump limit
when the solver explores the search tree.  This change simplifies the log
processing step and allows us to skip creating the log when we don't need to
print it, though that's not implemented in this commit.

This commit changes the meaning of the backjump count slightly.  Previously the
solver counted a backjump every time it backtracked and added at least one
variable to the conflict set.  Now it counts a backjump every time it unions all
conflict sets at the current level, even if the union is the same as the
conflict set from the level below.  That means that more steps count as
backjumps, and the solver can reach the backjump limit slightly earlier.

This commit also changes the log messages.  Previously, the solver sometimes
omitted the last 'Backjump' message, but this commit always prints a 'Backjump'
message when the solver fails.  Additionally, the level numbers for some
'Backjump' messages changed to match the levels where the conflict sets were
unioned.