Use RetryLog to avoid an O(N) update to the end of the solver log.
This commit delays converting the solver's log from a RetryLog (See #3674) to a
Progress as long as possible, until immediately before the call to
'showMessages', which needs to pattern match on the log messages.  This
maximizes the number of operations on the end of the log that are done in O(1)
time instead of O(N) time.