Make cabal check warn about missing directories in globs
This also significantly improves the error when trying to refer to
missing directories, hopefully making it clear that it's coming from
Cabal. #5318 and snowleopard/hadrian#634 are two bugs which manifested
as Cabal trying to glob in a non-existent directory and both took some
debugging because of the obscurity of the error.