Do register monolithic packages if they contain a buildable pub-lib (#5604)
As the case of `cpphs-1.20.8` has shown in #5379, in cases when a
monolithic package gets installed due to its executable components
(i.e. `exe:cpphs`) into the store we *have* to register if there's a
buildable public library (i.e. `lib:cpphs`) that was built and installed
into the same store folder as otherwise this will cause build failures
once a target actually depends on `lib:cpphs`

This ought to fix #5379