Link to 'developing packages' from new-build docs.
This page is often surfaces above the "Developing Quickstart" section in search results but a beginner ending up here may not know how to set up their package before using the commands here. This adds a link to the section in the docs that walks through how to set up a package so they don't need to go looking for it themselves.