validate.sh: Allow using short-version ghc binaries, like ghc-8.6
Previously this would work until CABAL_TESTSUITE_BDIR was needed, at
which point it would fail because it looked for things under
.../ghc-8.6/... while cabal v2-build actually put them in
.../ghc-8.6.5/...

Hard-coding "ghc" is unfortunate, but I didn't know if there was an easy
way to avoid it.  I also didn't want to over-complicate it, until it was
necessary.

validate.sh isn't tested in CI so...
[ci skip]