Solver: Support dependencies on sub-libraries (issue #6039).
This commit tracks dependencies on sub-libraries by extending the functionality
for tracking executables that was added in
640e526ce0e2514a1c7bc982da76a0d9faf3b6df.

It also starts adding support for library visibility, though it currently only
works for source packages.  There is a TODO for handling installed packages.

This commit handles visibility similarly to the way that the buildable field is
handled currently.  It only checks whether a component is made private by the
current environment and flag constraints at the start of dependency solving.
This means that the solver can treat a component as visible when the visibility
is controlled by an automatic flag, and the build can fail later, depending on
the value that is chosen for the flag.

Fixes #6038.