cabal-install: Rewrite bootstrap.sh with Python3
The plan: A JSON file (generated with bootstrap.py --extract-plan)
containing dependency and source information is built from plan.json.
This is then used to fetch and build the bootstrapping dependencies.
Since this is only to be used for bootstrapping on new architectures, we
advise the users to only use the resulting cabal-install executable to
build a proper cabal-install.