Re #7331: factor error message decoration in variants of `die`
The existing code had quite some duplication, which makes it more
brittle and harder to understand.

I factored out atomic modifications of the error message that will
eventually be raised by 'die' and friends.  These factors can now be
combined in a pipeline.

The factors that require IO are:

  - prefix with program name (e.g. `cabal`)
  - annotate with time stamp and meta-data

The lens `ioeErrorString` and its pure compagnion
`ioeModifyErrorString` should really be in `base:System.IO.Error`,
but they aren't, so I define them here.