Remove any remnant of the 'world' file
In v1-install, 'world' was used to trace what packages have been
installed, and re-install on-demand everything listed in it. However, in
v2-install, this is no longer needed and this file leads to code bloat.
Additionally, 'world' code-path is probably not up-to-date, remove it instead of
having partial features.